sfojf

+++ sagittal_brain.py
@@ -12,7 +12,7 @@ def run_averages(file_input='brain_sample.csv', file_output='brain_average.csv')
     The result is the average for each sagittal/horizontal plane (rows)
     """
     # Open the file to analyse
-    planes = np.loadtxt(file_input, dtype=int,  delimiter=',')
+    planes = np.loadtxt(file_input, dtype=float,  delimiter=',')
 
     # Calculates the averages through the sagittal/horizontal planes
     # and makes it as a row vector
